requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-docx==1.1.2
requests==2.32.3
openai==1.35.3
python-multipart==0.0.9
aiofiles==23.2.1
python-dotenv==1.0.1
